Issue #14937: Migrated section 3.4.1 Exactly one top-level class declaration to chapter wise testing
