+++ src/it/resources/com/google/checkstyle/test/chapter3filestructure/rule341onetoplevel/InputOneTopLevelClassBad1.java
@@ -1,5 +1,7 @@
 package com.google.checkstyle.test.chapter3filestructure.rule341onetoplevel;
 
 class InputOneTopLevelClassBad1 {} //ok
-enum FooEnum {} // warn
+// violation below 'Top-level class FooEnum has to reside in its own source file.'
+enum FooEnum {}
+// violation below 'Top-level class FooAt has to reside in its own source file.'
 @interface FooAt {} // warn
